pyproject.toml
[project]
name = "systems"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.11.0",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=22.6",
    "supabase>=2.15.0",
]
